requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.2

# Movie Search Dependencies
cinemagoer>=2023.5.1
fuzzywuzzy[speedup]>=0.18.0
python-Levenshtein>=0.21.0

# Image Processing (for wheel spinning)
Pillow>=10.0.0
imageio>=2.31.1
numpy>=1.24.3

# Database Dependencies
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Environment and Configuration
python-dotenv>=1.0.0

# Development Dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For better performance
asyncpg>=0.28.0
uvloop>=0.17.0; sys_platform != "win32"